KMC-6 - Got all the necessary plumbing in place for a Source connector. I have no idea at this point if any of this works.
KMC-6 - This gets the Source connector to actually work. However, I doubt I'm using DMSDK correctly, and there are certainly some basic improvements to be made.

KMC-6 - A little cleanup.

KMC-6 - A little cleanup.

KMC-6 - A rough prototype of a source using the ML Java client without the DMSDK. This version deletes a file after the Kafka message is committed. I need to change this to an option.

KMC-6 - Changed the document delete on Kafka commit to simply changing the collection.
Also did a little more work on creating the Connector Archive.

KMC-6 - This is a prototype of a Source task to pull records from MarkLogic and post them to a Kafka topic. Note that only the MarkLogicClientSourceTask is functioning as expected. The DMSDK version is not ready yet.